+++ qiskit_nature/transformers/seniority_zero_transformer.py
@@ -19,7 +19,7 @@ from ..drivers import QMolecule
 class SeniorityZeroTransformer(BaseTransformer):
     """The Seniority-Zero transformer."""
 
-    def transform(self, q_molecule: QMolecule) -> QMolecule:  # type: ignore
+    def transform(self, q_molecule: QMolecule) -> QMolecule:
         """Transforms the given `QMolecule` into a seniority-zero (i.e.
         restricted-spin) variant.
